Halytics.Metric

Description

...

Documentation

>>> :set -XDataKinds
>>> :set -XTypeOperators
>>> :set -XFlexibleContexts
>>> :set -XKindSignatures
>>> import Safe (maximumMay, minimumMay)

data All :: *

All will simply give back all the entries collected. This should mostly be used for testing.

>>> let monitor = generate :: Monitor All
>>> result (notifyMany monitor [1.0,2.0,3.0]) :: [Double]
[1.0,2.0,3.0]

The following property should hold:

result (notifyMany (generate :: Monitor All) xs) == xs

data Max

Max will result in the largest entry collected so far. If no entry was collected so far, results in Nothing. If any entry was collected, results in Just the maximum.

>>> let monitor = generate :: Monitor Max
>>> result (notifyMany monitor [1.0,2.0,3.0]) :: Maybe Double
Just 3.0
>>> result (notifyMany monitor []) :: Maybe Double
Nothing

The following property should hold:

result (notifyMany (generate :: Monitor Max) xs) == maximumMay xs

data Min

Min will result in the smallest entry collected so far. If no entry was collected so far, results in Nothing. If any entry was collected, results in Just the minimum.

>>> let monitor = generate :: Monitor Min
>>> result (notifyMany monitor [1.0,2.0,3.0]) :: Maybe Double
Just 1.0
>>> result (notifyMany monitor []) :: Maybe Double
Nothing

The following property should hold:

result (notifyMany (generate :: Monitor Min) xs) == minimumMay xs

data Every :: Nat -> * -> *

Every will feed another metric every nth element collected. Every will feed the metric the first, n+1 th, 2n+1 th, ... collected elements.

>>> let monitor = generate :: Monitor (All |^ Every 2)
>>> result (notifyMany monitor [1.0,2.0,3.0,2.0,1.0]) :: [Double]
[1.0,3.0,1.0]

data Last :: Nat -> * -> *

Last will feed another metric the last n elements collected.

>>> let monitor = generate :: Monitor (All |^ Last 2)
>>> result (notifyMany monitor [1.0,2.0,3.0,2.0,1.0]) :: [Double]
[2.0,1.0]

data PeriodOf :: Nat -> * -> *

PeriodOf will feed another metric with chunks of n elements. result returns the list of the results (one for each chunk).

>>> let monitor = generate :: Monitor (All |^ PeriodOf 2)
>>> result (notifyMany monitor [1.0,2.0,3.0,2.0,1.0]) :: [[Double]]
[[1.0,2.0],[3.0,2.0],[1.0]]